requirements.txt
Django==6.0.1
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.2,<6
django-redis>=5.2,<6
redis>=4.5,<5
argon2-cffi>=21.1,<23
psycopg2-binary>=2.9,<3
django-environ>=0.9.0
whitenoise>=6.4.0

# Optional / dev
ipython
pytest
pytest-django
gunicorn>=20.1.0
